feat(audittrail): add request input to AuditTrail
